improvements on storage file handling
